Initial commit for decomposing the the monolithic builder
