+++ Model/Product/ProductProvider.php
@@ -1,0 +1,14 @@
+<?php
+
+
+namespace Nosto\Tagging\Model\Product;
+
+
+use Magento\Catalog\Model\Product;
+use Magento\Store\Model\Store;
+use Nosto\Object\Product\Product as NostoProduct;
+
+interface ProductProvider
+{
+    function addData(Product $product, Store $store, NostoProduct $nostoProduct);
+}
